test(integration): Fix integration test and make use of new OpKind enum
