+++ integration-tests/package.json
@@ -7,7 +7,8 @@
   "dependencies": {
     "@taquito/local-forging": "^6.0.3-beta.1",
     "@taquito/signer": "^6.0.3-beta.1",
-    "@taquito/taquito": "^6.0.3-beta.1"
+    "@taquito/taquito": "^6.0.3-beta.1",
+    "@taquito/utils": "^6.0.3-beta.1"
   },
   "private": true,
   "jest": {
