+++ packages/taquito-rpc/src/types.ts
@@ -1,4 +1,5 @@
 import BigNumber from 'bignumber.js';
+import { OpKind } from './opkind';
 
 export type BalanceResponse = BigNumber;
 export type StorageResponse = ScriptedContracts['storage'];
@@ -26,20 +27,6 @@ interface Frozenbalancebycycle {
   deposit: BigNumber;
   fees: BigNumber;
   rewards: BigNumber;
-}
-
-export enum OpKind {
-  ORIGINATION = 'origination',
-  DELEGATION = 'delegation',
-  REVEAL = 'reveal',
-  TRANSACTION = 'transaction',
-  ACTIVATION = 'activate_account',
-  ENDORSEMENT = 'endorsement',
-  SEED_NONCE_REVELATION = 'seed_nonce_revelation',
-  DOUBLE_ENDORSEMENT_EVIDENCE = 'double_endorsement_evidence',
-  DOUBLE_BAKING_EVIDENCE = 'double_baking_evidence',
-  PROPOSALS = 'proposals',
-  BALLOT = 'ballot',
 }
 
 export type BigMapKey = { key: { [key: string]: string }; type: { prim: string } };
